final5

+++ Library/library-elchin-emirov/src/main/java/az/developia/libraryelchinemirov/dao/StudentDAO.java
@@ -1,11 +1,12 @@
 package az.developia.libraryelchinemirov.dao;
 
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
 import az.developia.libraryelchinemirov.entity.StudentEntity;
 
 @Repository
-public interface StudentDAO extends JpaRepository<StudentEntity, Integer>{
-	
+public interface StudentDAO extends JpaRepository<StudentEntity, Long> {
 
 }
